Divide account handlers into modules
