+++ server/api/account/handlers/index.js
@@ -1,0 +1,6 @@
+exports.changePassword = require('./changePassword');
+exports.login = require('./login');
+exports.resetPassword = require('./resetPassword');
+exports.sendInviteEmail = require('./sendInviteEmail');
+exports.sendResetPasswordEmail = require('./sendResetPasswordEmail');
+exports.signup = require('./signup');
